Add support for decompression of HTTP responses
If the module configuration specifies the "compression" option, with the
expected compression algorithm, blackbox_exporter will try to decompress
the response. If the response is not compressed using the specified
algorithm, the probe will fail.

It validates that the "Accept-Encoding" header is either absent, or that
it specifies the same algorithm as the "compression" option.

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>